requirements.txt
requests>=2.31.0
Pillow>=10.0.0
azure-cognitiveservices-speech>=1.31.0
moviepy==1.0.3
decorator>=4.0.2
numpy>=1.24.0
python-dotenv>=1.0.0
imageio-ffmpeg>=0.4.3
ollama>=0.1.0
pydub>=0.25.1
